Crate srml_metadata

Decodable variant of the RuntimeMetadata.

This really doesn't belong here, but is necessary for the moment. In the future it should be removed entirely to an external module for shimming on to the codec-encoded metadata.

Structs

DefaultByteGetter
Wrapper over dyn pointer for accessing a cached once byte value.

EventMetadata
All the metadata about a event.

FnEncode
Newtype wrapper for support encoding functions (actual the result of the function).

FunctionArgumentMetadata
All the metadata about a function argument.

FunctionMetadata
All the metadata about a function.

ModuleMetadata
All metadata about an runtime module.

OuterEventMetadata
All the metadata about an outer event.

RuntimeMetadataPrefixed
Metadata prefixed by a u32 for reserved usage

RuntimeMetadataV5
The metadata of a runtime.

StorageFunctionMetadata
All the metadata about a storage function.

StorageMetadata
All the metadata about a storage.

Enums

DecodeDifferent
A type that decodes to a different type than it encodes. The user needs to make sure that both types use the same encoding.

RuntimeMetadata
The metadata of a runtime. The version ID encoded/decoded through the enum nature of RuntimeMetadata.

RuntimeMetadataDeprecated
Enum that should fail.

StorageFunctionModifier
A storage function modifier.

StorageFunctionType
A storage function type.

StorageHasher
Hasher used by storage maps

Constants

META_RESERVED
Curent prefix of metadata

Traits

DefaultByte
A technical trait to store lazy initiated vec value as static dyn pointer.

Type Definitions

ByteGetter
Decode different for static lazy initiated byte value.

DecodeDifferentArray